requirements.txt
dj-static==0.0.6
Django==1.9.5
djangorestframework==3.3.3
Pillow==3.2.0
psycopg2==2.6.1
static3==0.6.1
